refactor : Table -> Order 도메인으로 변경
