+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -1,6 +1,5 @@
 package kitchenpos.order.domain;
 
-import kitchenpos.table.domain.OrderTable;
 import org.springframework.util.CollectionUtils;
 
 import javax.persistence.*;
