+++ src/main/java/kitchenpos/order/domain/OrderTableRepository.java
@@ -1,9 +1,7 @@
-package kitchenpos.table.domain;
+package kitchenpos.order.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
 
 @Repository
 public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
