+++ src/main/java/kitchenpos/order/dto/OrderTableRequest.java
@@ -1,6 +1,6 @@
-package kitchenpos.table.dto;
+package kitchenpos.order.dto;
 
-import kitchenpos.table.domain.OrderTable;
+import kitchenpos.order.domain.OrderTable;
 
 public class OrderTableRequest {
     private int numberOfGuests;
