+++ src/main/java/kitchenpos/order/dto/TableGroupRequest.java
@@ -1,7 +1,7 @@
-package kitchenpos.table.dto;
+package kitchenpos.order.dto;
 
-import kitchenpos.table.domain.OrderTable;
-import kitchenpos.table.domain.TableGroup;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.TableGroup;
 
 import java.util.List;
